Added key authentication to the API, app, & website

+++ server/index.js
@@ -28,6 +28,11 @@ function generateConfirmationCode() {
   return crypto.randomBytes(20).toString("hex");
 }
 
+function isSigned(key) {
+  if (key == process.env.ACCESS_KEY) return 1;
+  return 0;
+}
+
 app.use(cors());
 app.use(express.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -54,6 +59,7 @@ app.get("/mailtest", (req, res) => {
 //reports database insert
 app.post("/insert", async (req, res) => {
   const conf_code = generateConfirmationCode();
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   try {
     const coords = await getCoordinates(
@@ -221,6 +227,7 @@ app.get("/confirm-email", (req, res) => {
 app.post("/sendCode", (req, res) => {
   const email = req.body.email;
   const code = req.body.code;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   //send email with code
   const messagebody =
@@ -252,6 +259,7 @@ app.post("/bk_insert", (req, res) => {
   const city = req.body.city;
   const zip = req.body.zip;
   const pass = req.body.pass;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   bcrypt.genSalt(10, function (err, salt) {
     if (err) return console.log(err);
@@ -283,6 +291,7 @@ app.post("/bk_update", (req, res) => {
   const city = req.body.city;
   const zip = req.body.zip;
   const bk_id = req.body.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlUPDATE =
     "UPDATE beekeepers SET fname = ?, lname = ?, phone_no = ?, address = ?, city = ?, zip = ? WHERE bk_id = ?;";
@@ -316,6 +325,8 @@ app.post("/bk_qualif_update", (req, res) => {
   const ladder = req.body.ladder; // LADDER IS OF TYPE INT
   const mechanical_lift = req.body.mechanical_lift;
   const bk_id = req.body.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
+
   const sqlUPDATE =
     "UPDATE qualifications SET ground_swarms = ?, valve_or_water_main = ?, shrubs = ?, low_tree = ?, mid_tree = ?, tall_tree = ?, fences = ?, low_structure = ?, mid_structure = ?, chimney = ?, interior = ?, cut_or_trap_out = ?, traffic_accidents = ?, bucket_w_pole = ?, ladder = ?, mechanical_lift = ? WHERE bk_id = ?;";
   db.query(
@@ -352,6 +363,7 @@ app.post("/bk_qualif_update", (req, res) => {
 app.post("/bk_pass_update", (req, res) => {
   const pass = req.body.pass;
   const bk_id = req.body.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   bcrypt.genSalt(10, function (err, salt) {
     if (err) return console.log(err);
@@ -374,6 +386,8 @@ app.post("/bk_pass_update", (req, res) => {
 app.post("/complete_report", (req, res) => {
   const r_id = req.body.r_id;
   //const active = req.body.active;
+
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   //mark report as complete
   const sqlUpdate = "UPDATE reports SET complete = true WHERE r_id = ?;";
@@ -423,6 +437,7 @@ app.post("/complete_report", (req, res) => {
 app.post("/abandon_report", (req, res) => {
   const r_id = req.body.r_id;
   console.log("Abandoning report #" + r_id);
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlDelete = "DELETE FROM active_reports WHERE r_id = ?;";
   db.query(sqlDelete, [r_id], (err, result) => {
@@ -443,6 +458,7 @@ app.post("/abandon_report", (req, res) => {
 app.post("/remove_report", (req, res) => {
   const r_id = req.body.r_id;
   const reason = req.body.reason;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
   console.log("Removing report #" + r_id);
   console.log("Reason: " + reason);
 
@@ -472,6 +488,7 @@ app.post("/remove_report", (req, res) => {
 app.use("/claim_report", (req, res) => {
   const r_id = req.body.r_id;
   const bk_id = req.body.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlInsert = "INSERT INTO active_reports (bk_id, r_id) VALUES (?, ?)";
   db.query(sqlInsert, [bk_id, r_id], (err, result) => {
@@ -522,6 +539,7 @@ app.use("/claim_report", (req, res) => {
 // Fetches a user email and bk_id from the Beekeepers table
 app.get("/bk_user", (req, res) => {
   const email = req.query.email;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlQuery = "SELECT email, bk_id FROM beekeepers WHERE email = ?;";
   db.query(sqlQuery, [email], (err, result) => {
@@ -534,6 +552,7 @@ app.get("/bk_user", (req, res) => {
 //not being used as of now
 app.get("/bk_pass", (req, res) => {
   const bk_id = req.query.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlQuery = "SELECT email, pass FROM beekeepers WHERE bk_id = ?";
   db.query(sqlQuery, [bk_id], (err, result) => {
@@ -552,6 +571,7 @@ app.get("/bk_get", (req, res) => {
   //MUST use query when making a get request to the database!
   const email = req.query.email;
   const pass = req.query.pass;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   //select the bk using the email then check the hashed password
   const sqlQuery = "SELECT * FROM beekeepers WHERE email = ?";
@@ -576,6 +596,7 @@ app.get("/bk_get", (req, res) => {
 
 app.get("/bk_getUser", (req, res) => {
   const bk_id = req.query.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
   const sqlQuery = "SELECT * FROM beekeepers WHERE bk_id = ?;";
   db.query(sqlQuery, [bk_id], (err, result) => {
     if (err) return res.status(500).send(err.message);
@@ -585,6 +606,7 @@ app.get("/bk_getUser", (req, res) => {
 
 // Fetch bee reports to display on the app
 app.get("/bk_appReports", (req, res) => {
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
   const sqlQuery =
     "SELECT * FROM reports WHERE active = false AND confirmed = true;";
   db.query(sqlQuery, (err, result) => {
@@ -596,6 +618,7 @@ app.get("/bk_appReports", (req, res) => {
 // Fetches reports a beekeeper has claimed for MyReports for a report to show here active must be true
 app.get("/bk_claimedReports", (req, res) => {
   const bk_id = req.query.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlQuery =
     "SELECT * FROM reports JOIN active_reports ON reports.r_id = active_reports.r_id WHERE reports.active = true AND active_reports.bk_id = ?;";
@@ -609,6 +632,7 @@ app.get("/bk_claimedReports", (req, res) => {
 // Fetches reports a beekeeper has completed for MyReports
 app.get("/bk_completedReports", (req, res) => {
   const bk_id = req.query.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlQuery = "SELECT * FROM report_archive WHERE bk_id = ?";
   db.query(sqlQuery, [bk_id], (err, result) => {
@@ -646,6 +670,7 @@ app.get("/debug-report", (req, res) => {
 //fetches all info for a specifc report
 app.get("/report_data", (req, res) => {
   const r_id = req.query.r_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlQuery = "SELECT * FROM reports WHERE r_id = ?;";
   db.query(sqlQuery, [r_id], (err, result) => {
@@ -656,6 +681,7 @@ app.get("/report_data", (req, res) => {
 
 app.get("/bk_getFull", (req, res) => {
   const bk_id = req.query.bk_id;
+  if (!isSigned(req.body.key)) return res.status(400).send("Bad key.");
 
   const sqlQuery =
     "select * from beekeepers NATURAL JOIN qualifications WHERE bk_id = ?;";
